+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [2.22.0] - 2021-09-14
+### Added
+- Added available options in menu dropdown
+### Fixed
+- Fixed track file handler
+
 ## [2.21.0] - 2021-09-14
 ### Added
 - Unit tests added for utils, logger and commands_handler